+++ src/batou/lib/git.py
@@ -21,19 +21,25 @@ def ensure_empty_directory(path):
         ensure_path_nonexistent(os.path.join(path, file))
 
 
+def exactly_one(*args):
+    return sum(map(bool, args)) == 1
+
+
 class Clone(Component):
 
     namevar = "url"
     target = "."
     revision = None
     branch = None
+    tag = None
     vcs_update = True
     clobber = False
 
     def configure(self):
-        if (not self.revision_or_branch) or (self.revision and self.branch):
+        if not exactly_one(self.revision, self.branch, self.tag):
             raise ValueError(
-                "Clone(%s) needs exactly one of revision or branch" % self.url
+                "Clone(%s) needs exactly one of revision, branch or tag"
+                % self.url
             )
         self.target = self.map(self.target)
         self += Directory(self.target)
@@ -92,10 +98,18 @@ class Clone(Component):
                 or self.has_incoming_changesets()
             ):
                 raise UpdateNeeded()
+            if self.tag and (
+                self.current_tag() != self.tag or self.has_incoming_changesets()
+            ):
+                raise UpdateNeeded()
 
-    @property
-    def revision_or_branch(self):
-        return self.revision or self.branch
+    def current_tag(self):
+        try:
+            with self.chdir(self.target):
+                stdout, stderr = self.cmd("LANG=C git describe --tags")
+        except RuntimeError:
+            return None
+        return stdout.strip()
 
     def current_revision(self):
         try:
@@ -147,6 +161,8 @@ class Clone(Component):
                 self.cmd(
                     self.expand("git reset --hard origin/{{component.branch}}")
                 )
+            elif self.tag:
+                self.cmd(self.expand("git reset --hard {{component.tag}}"))
             else:
                 self.cmd(self.expand("git reset --hard {{component.revision}}"))